+++ protein.h
@@ -1,5 +1,3 @@
-/* Contact: Toni Westbrook <anthonyw@wildcats.unh.edu> */
-
 #ifndef PROTEIN_H_
 #define PROTEIN_H_
 
@@ -11,10 +9,11 @@
 #define OUTPUT_TYPE_UNIPROT_SIMPLE 0
 #define OUTPUT_TYPE_UNIPROT_FULL   1
 
-#define ALIGN_FLAG_BRUTE_ORF   0x0001
+#define ALIGN_FLAG_BRUTE_ORF  0x0001
 #define ALIGN_FLAG_GEN_NT     0x0002
 #define ALIGN_FLAG_KEEP_PRO   0x0004
 #define ALIGN_FLAG_ADJUST_ORF 0x0008
+#define ALIGN_FLAG_MANUAL_PRO 0x0010
 
 extern unsigned char codon_aa_hash[64];